refactor: clean up dead code and reorganize API layer
- Delete unused `$lib/stores/` directory (auth.ts, links.ts)
- Move `fetchUrlTitle()` to API layer (`frontend/src/lib/api/metadata.ts`)
- Move `debounce()` utility to proper location (`frontend/src/lib/utils/helpers.ts`)
- Add `fetchUrlTitle()` method to `linksApi` for cleaner API calls
- Replace `apiClient.request()` with `apiClient.put()` in `linksApi.update()`
- Update `LinkModal.svelte` to use new API locations
- Add `put()` method to `ApiClient` class for consistent API patterns
