+++ frontend/src/lib/api/metadata.ts
@@ -1,4 +1,4 @@
-import { apiClient } from "$lib/api/client";
+import { apiClient } from "./client";
 
 interface FetchTitleResponse {
   title: string | null;
@@ -31,23 +31,4 @@ export async function fetchUrlTitle(url: string): Promise<string | null> {
     console.debug("Failed to fetch URL title:", error);
     return null;
   }
-}
-
-/**
- * Debounce function to limit how often a function is called
- * @param func - The function to debounce
- * @param delay - Delay in milliseconds
- * @returns Debounced function
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function debounce<T extends (...args: any[]) => any>(
-  func: T,
-  delay: number
-): (...args: Parameters<T>) => void {
-  let timeoutId: ReturnType<typeof setTimeout>;
-
-  return (...args: Parameters<T>) => {
-    clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => func(...args), delay);
-  };
 }
